requirements-vercel.txt
openai>=1.12.0
anthropic>=0.19.0
google-generativeai>=0.5.0
dnspython>=2.4.0
beautifulsoup4>=4.13.0
colorama>=0.4.6
playwright>=1.40.0
PyYAML>=6.0.0
requests>=2.30.0
tiktoken>=0.5.0
flask>=2.2.0
markdown>=3.4.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
# Explicitly specify Flask-CORS rather than using 'flask_cors' format
Flask-CORS>=4.0.0
# Use Vercel Analytics (@vercel/analytics)
vercel-analytics>=0.1.1
